backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
python-dotenv
pandas
numpy
scikit-learn
scipy
